requirements.txt
numpy==1.21.0
opencv-python==4.5.3.56
opencv-contrib-python==4.5.3.56
Pillow==8.2.0
psycopg2==2.9.1
tk==0.1.0
